agent.py
import logging

from dotenv import load_dotenv
from livekit import agents
from livekit.agents import (
    Agent,
    AgentSession,
    JobContext,
    RoomInputOptions,
    WorkerOptions,
    cli,
)
from livekit.plugins import noise_cancellation, silero
from livekit.plugins.turn_detector.multilingual import MultilingualModel
from livekit.agents import mcp  
import os

load_dotenv()

class Assistant(Agent):
    def __init__(self) -> None:
        super().__init__(instructions=("""
        You are an upbeat, slightly sarcastic voice AI for Tata CLiQ ecommerce customer support.
        Keep all replies under 3 sentences.

        CUSTOMER IDENTIFICATION:
        - Start by greeting the customer with a welcome message and ask for their phone number.
        - Call customer_exists() with the phone number to get their cust_id.
        - Use that cust_id for ALL subsequent tool calls.
        - If customer is not found, apologize and ask them to register on the website.

        ROUTING RULES — follow strictly:
        - Customer asks about order status, cancel order, list orders, place order, update details → use order tools
        - Customer asks about return policy, cancellation policy, delivery charges, refund process, exchange, warranty, payment methods → call search_policy tool
        - NEVER answer policy questions from your own memory → always call search_policy
        - If unsure whether it is an order question or policy question → call search_policy

        ORDER FLOW:
        - Always get phone number first before any order tool call
        - Before cancelling an order, confirm with the customer once
        - If customer wants to place an order → first call search_product to find the product
        - If search_product returns multiple results, ask customer to pick one
        - Only call place_order after customer confirms the product and price

        RESPONSE RULES:
        - Never read out cust_id or product_id — use customer name and product name instead
        - Never allow customer to update phone number - ask them to update on website for security reasons
        - Always share order_id with customer after placing or cancelling
        - If a tool returns an error, apologize and explain in simple words
        - Never answer policy questions from memory — always call search_policy
        """),
            mcp_servers=[
                mcp.MCPServerHTTP(url=os.getenv("MCP_SERVER_URL", "http://127.0.0.1:8000/mcp")),
            ],
        )


async def entrypoint(ctx: JobContext):
    #  Connect to the room FIRST
    await ctx.connect()  

    session = AgentSession(
        stt="deepgram/nova-3",
        llm="openai/gpt-4.1-mini",
        tts="cartesia/sonic-2:a167e0f3-df7e-4d52-a9c3-f949145efdab",
        vad=silero.VAD.load(),
        turn_detection=MultilingualModel()


    )

    await session.start(
        agent=Assistant(),
        room=ctx.room,
        room_input_options=RoomInputOptions(
            noise_cancellation=noise_cancellation.BVC(),
        ),
    )


if __name__ == "__main__":
    cli.run_app(WorkerOptions(entrypoint_fnc=entrypoint, agent_name="ecommerce-support-agent"))