pyproject.toml
[project]
name = "ecommerce-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "fastmcp>=3.1.1",
    "livekit-agents[mcp,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
]
